Move the static worker pool instance to style.js
This is still not great, but it's better than having it in Dispatcher.
Ultimately the right place to put it is at the root level, i.e.
mapbox-gl.js, and then pass it through, but that will be a noisy change
due the number of 'new Map()' and 'new Style()' in the unit tests